requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# ML dependencies
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0

# Web interface
streamlit==1.28.1
streamlit-image-coordinates==0.1.5

# Data processing
pycocotools==2.0.7
pyyaml==6.0.1
pandas==2.1.3
scikit-learn==1.3.2

# Utilities
tqdm==4.66.1
requests==2.31.0

